+++ src/main/test/modules/date-picker/bordered/bordered.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ex-bordered',
+  templateUrl: './bordered.component.html',
+  styleUrls: ['./bordered.component.scss']
+})
+export class ExBorderedComponent {}