+++ src/services/config.service.ts
@@ -10,7 +10,7 @@ export class ConfigService {
   langs = ['zh_CN', 'en_US'];
   cacheLangs: { [lang: string]: XI18nProperty } = {};
   versions: string[] = [];
-  version = '15.0.1';
+  version = '15.0.3';
   navName = 'NG-NEST';
 
   get lang() {